Using UnreadCountLogic
